Visual Changes
Readability stuff. Shoutout to Ahmet for asking about that. This whole time I was thining about how to make the text more visable and finding a darker video… forgot that css can do tricks.

Dropped opacity for background video and images, and placed a black background behind the body and the div containing the animal images. Incrased font weights on card titles, animal name and city, to be more readable.